+++ src/store/modules/menu/constant.ts
@@ -1,0 +1,11 @@
+export const pickRouteRecordNormalizedConstant = [
+  'redirect',
+  'breadcrumbLabel',
+  'children',
+  'key',
+  'meta',
+  'name',
+  'path',
+  'show',
+  'fullPath',
+]